Unify test-data loaders under test/<db>/load_test_data.* and prune dead tables
Every load-into-an-existing-instance dialect now follows one convention:
test/<db>/load_test_data.sh wraps a .ts/.sql payload, absorbing the runner,
working directory, and env handling behind a uniform entrypoint. The DuckDB
loader moves out of scripts/ to sit beside the others, and the dead
snowflake_tables.sh generator is deleted (long superseded by the checked-in
test/snowflake/load_test_data.sql).

Prune tables that no test queries: numbers, words, words_bigger (loaded by
DuckDB/Snowflake/MSSQL) and flights_partitioned (Snowflake-only). Every
dialect now loads the identical 8-table malloytest set, ending the drift
where each loader hand-maintained a different list. The one test that read
words.parquet was exercising DuckDB's file-path table grammar, not the
words table, so it is repointed at a small FAA parquet (carriers).

Also removes a few machine-specific paths that had leaked into repo
comments and docs.
