+++ package.json
@@ -68,7 +68,7 @@
     "consumer-canary": "npm run consumer-canary:bundle && npm run consumer-canary:jest",
     "consumer-canary:bundle": "node test/consumer-canary/bundle-check.mjs",
     "consumer-canary:jest": "jest --config test/consumer-canary/jest.config.ts",
-    "build-duckdb-db": "ts-node scripts/build_duckdb_test_database",
+    "build-duckdb-db": "sh test/duckdb/load_test_data.sh",
     "setup-new-checkout": "scripts/setup-new-checkout.sh",
     "ping-db": "ts-node ./test/bin/ping_db.ts"
   },
