+++ test/data/duckdb/README.md
@@ -1,4 +1,4 @@
 # Building test database for duckdb
 
-1. At top-level, run `npx ts-node scripts/build_duckdb_test_database.ts`
+1. At top-level, run `sh test/duckdb/load_test_data.sh`
 2. A file called `duckdb_test.db` should be created in this folder
